CASMPET-4243: Pass in the database name
The database name tells this tool what database is being backed
up and from this it can calculate the pod name and the names of
the Secrets that patroni creates that need to be backed up.

The database name is used as the base for the keys in the bucket
for the objects that contain the backup.