CASMPET-4239: Stream pg_dump directly to s3
The script was putting the pg_dump output into a file and then
sending that file to s3. This would likely cause problems with
large databases because the disk would fill up or it would just
take a long time to write the file and read it out again.

The fix here is to stream the pg_dump output to s3.