CASMPET-4232: Backup a replica if it's available
Rather than run the backup on the leader this checks to see if
there's a replica that's not far behind and uses that rather than
the leader. This puts less load on the leader which already has
more load on it due to being the target of all updates and doing
replication.